webapp/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
pandas>=2.0.0
pillow>=10.0.0
plotly>=5.18.0
